requirements.txt
tkinter
websockets
Pillow

